Implements insurances classes

+++ WebClientApi/src/main/java/com/fernando/Controllers/ClientPJController.java
@@ -33,7 +33,7 @@ public class ClientPJController {
 	
 	//FindById Controller
 	@GetMapping(value = "/{id}",produces = MediaType.APPLICATION_JSON_VALUE)
-	public ClientPJ findById(@PathVariable(value = "id") Integer id) {
+	public ClientPJ findById(@PathVariable(value = "id") Long id) {
 		return service.findById(id);
 	}
 	
@@ -53,7 +53,7 @@ public class ClientPJController {
 	
 	//Delete Controller
 	@DeleteMapping(value = "/{id}")
-	public ResponseEntity<?> delete(@PathVariable (value = "id") Integer id) {
+	public ResponseEntity<?> delete(@PathVariable (value = "id") Long id) {
 		service.delete(id);
 		return ResponseEntity.noContent().build();
 	}
